sync

+++ src/system.cpp
@@ -56,5 +56,5 @@ int System::TotalProcesses() {
 
 // DONE: Return the number of seconds since the system started running
 long int System::UpTime() { 
-    return LinuxParser::UpTime();
+    return (long int) LinuxParser::UpTime();
  }
